Add files via upload
It is a Python-based Flask API file that run at REST endpoint. This file is developed to compare the effect of running tests with an API similar to the serverless function that runs at an Endpoint. This file can be further developed and can be used in Other Serverless platforms that support python to code and Deploy the function